Bot checks which releases channels are registered before starting up
